0005786: System messages don't parse [DOMAIN] or [WEBSITE]
I tested and can confirm this happens when you enter [DOMAIN] and use Save and Activate. Using other scenario's I could not reproduce this issue.

Replacing [DOMAIN] with the domain gets done in defaultconfig.getConfig. This function is only called when building the page, as we do after 'Save'ing the page. After an 'Activate' the page is only stored. defaultconfig.getUserConfig is used when sending mail.

Added parsing of [DOMAIN\] and [WEBSITE\] to getUserConfig.

Parsing should be done only in one place. So maybe Later: Combine getConfig and this one?