0011163: SQL error on import: syntax error near 'data' since additional data is used
Fixed, code contains too many arrays with mappings and attributes IMHO. It works now but needs mayor redesign to be truly maintainable.

KNOWN ISSUE: When using test preview, system attribute mapping will be lost and needs to be selected twice. Annoying but no loss of functionality or reliability so I'll leave this for now.